ut test brakdown

+++ src/include/rdb/faccmemory.h
@@ -3,15 +3,21 @@
 #include "fainterface.h"
 
 namespace rdb {
-/**
- * @brief Object that implements storage interface in memory
- *
- * Type: MEMORY
- */
+
+/// @brief Definicja klasy implementującej dostęp danych przechowywanych w pamięci
+///
+/// Obiekt memoryFileAccessor powinien:
+/// - umożliwiać odczyt i zapis danych w pamięci
+/// - zapewniać trwałość danych w pamięci podczas działania programu
+/// - implementować interfejs FileAccessorInterface, aby umożliwić integrację z innymi komponentami systemu
+/// - być zoptymalizowany pod kątem wydajności, aby nie wprowadzać nadmiernych opóźnień w przetwarzaniu danych
+/// - zarządzać pamięcią w sposób efektywny, aby uniknąć wycieków pamięci
+/// - umożliwiać kontrolę ilości przechowywanych danych za pomocą mechanizmu retencji, który usuwa najstarsze dane po przekroczeniu określonego limitu
+
 struct memoryFileAccessor : public FileAccessorInterface {
   std::string filename_;
   const ssize_t recordSize_;
-  const int retentionSize_;   // Retention size for the records, if set to no_retention, no limit is applied
+  const size_t retentionSize_;  // Retention size for the records, if set to no_retention, no limit is applied
   int removed_count_ = 0;     // Count of removed records, used for retention management
   enum { no_retention = 0 };  // Default retention size if not specified
  public:
